[GHA] update integration test triggering branches
The current scheduling of integration test triggers on every push
instead of cron as it outlined in the yaml.

It turns out GHA doesn't support scheduling of jobs with multiple
branches. GHA's linter doesn't point out anything about this either.

We are going to Use GHA's matrix feature to work around its limitation.

Closes: #6021